+++ flatlaf-demo/src/main/java/com/formdev/flatlaf/demo/DemoFrame.java
@@ -18,6 +18,7 @@ package com.formdev.flatlaf.demo;
 
 import java.awt.*;
 import java.awt.event.*;
+import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -49,6 +50,7 @@ import com.formdev.flatlaf.extras.FlatSVGUtils;
 import com.formdev.flatlaf.util.ColorFunctions;
 import com.formdev.flatlaf.util.FontUtils;
 import com.formdev.flatlaf.util.LoggingFacade;
+import com.formdev.flatlaf.util.SystemFileChooser;
 import com.formdev.flatlaf.util.SystemInfo;
 import net.miginfocom.layout.ConstraintParser;
 import net.miginfocom.layout.LC;
@@ -170,6 +172,48 @@ class DemoFrame
 	private void saveAsActionPerformed() {
 		JFileChooser chooser = new JFileChooser();
 		chooser.showSaveDialog( this );
+	}
+
+	private void openSystemActionPerformed() {
+		SystemFileChooser chooser = new SystemFileChooser();
+		chooser.setMultiSelectionEnabled( true );
+		chooser.addChoosableFileFilter( new SystemFileChooser.FileNameExtensionFilter(
+			"Text Files", "txt", "md" ) );
+		chooser.addChoosableFileFilter( new SystemFileChooser.FileNameExtensionFilter(
+			"PDF Files", "pdf" ) );
+		chooser.addChoosableFileFilter( new SystemFileChooser.FileNameExtensionFilter(
+			"Archives", "zip", "tar", "jar", "7z" ) );
+
+		if( chooser.showOpenDialog( this ) != SystemFileChooser.APPROVE_OPTION )
+			return;
+
+		File[] files = chooser.getSelectedFiles();
+		System.out.println( Arrays.toString( files ).replace( ",", "\n" ) );
+	}
+
+	private void saveAsSystemActionPerformed() {
+		SystemFileChooser chooser = new SystemFileChooser();
+		chooser.addChoosableFileFilter( new SystemFileChooser.FileNameExtensionFilter(
+			"Text Files", "txt", "md" ) );
+		chooser.addChoosableFileFilter( new SystemFileChooser.FileNameExtensionFilter(
+			"Images", "png", "gif", "jpg" ) );
+
+		if( chooser.showSaveDialog( this ) != SystemFileChooser.APPROVE_OPTION )
+			return;
+
+		File file = chooser.getSelectedFile();
+		System.out.println( file );
+	}
+
+	private void selectFolderSystemActionPerformed() {
+		SystemFileChooser chooser = new SystemFileChooser();
+		chooser.setFileSelectionMode( SystemFileChooser.DIRECTORIES_ONLY );
+
+		if( chooser.showOpenDialog( this ) != SystemFileChooser.APPROVE_OPTION )
+			return;
+
+		File directory  = chooser.getSelectedFile();
+		System.out.println( directory  );
 	}
 
 	private void exitActionPerformed() {
@@ -508,6 +552,9 @@ class DemoFrame
 		JMenuItem newMenuItem = new JMenuItem();
 		JMenuItem openMenuItem = new JMenuItem();
 		JMenuItem saveAsMenuItem = new JMenuItem();
+		JMenuItem openSystemMenuItem = new JMenuItem();
+		JMenuItem saveAsSystemMenuItem = new JMenuItem();
+		JMenuItem selectFolderSystemMenuItem = new JMenuItem();
 		JMenuItem closeMenuItem = new JMenuItem();
 		exitMenuItem = new JMenuItem();
 		JMenu editMenu = new JMenu();
@@ -606,6 +653,25 @@ class DemoFrame
 				saveAsMenuItem.setMnemonic('S');
 				saveAsMenuItem.addActionListener(e -> saveAsActionPerformed());
 				fileMenu.add(saveAsMenuItem);
+				fileMenu.addSeparator();
+
+				//---- openSystemMenuItem ----
+				openSystemMenuItem.setText("Open (System)...");
+				openSystemMenuItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_O, Toolkit.getDefaultToolkit().getMenuShortcutKeyMask()|KeyEvent.SHIFT_DOWN_MASK));
+				openSystemMenuItem.addActionListener(e -> openSystemActionPerformed());
+				fileMenu.add(openSystemMenuItem);
+
+				//---- saveAsSystemMenuItem ----
+				saveAsSystemMenuItem.setText("Save As (System)...");
+				saveAsSystemMenuItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_S, Toolkit.getDefaultToolkit().getMenuShortcutKeyMask()|KeyEvent.SHIFT_DOWN_MASK));
+				saveAsSystemMenuItem.addActionListener(e -> saveAsSystemActionPerformed());
+				fileMenu.add(saveAsSystemMenuItem);
+
+				//---- selectFolderSystemMenuItem ----
+				selectFolderSystemMenuItem.setText("Select Folder (System)...");
+				selectFolderSystemMenuItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_F, Toolkit.getDefaultToolkit().getMenuShortcutKeyMask()|KeyEvent.SHIFT_DOWN_MASK));
+				selectFolderSystemMenuItem.addActionListener(e -> selectFolderSystemActionPerformed());
+				fileMenu.add(selectFolderSystemMenuItem);
 				fileMenu.addSeparator();
 
 				//---- closeMenuItem ----